Added STDP params, removed dependencies on Ne
Enabled class-level control over STDP behavior by adding
appropriate variables for the params and inserting switches
where previously the number of excitatory neurons had been the
only control over plasticity. Plasticity is now a property of the
class of the presynaptic neuron. Now there are variables for the
intercept and time constant for both LTP and LTD, as well as a
boolean for whether the neuron is plastic or not. Non-plastic
classes must have zero intercepts (but retain non-zero time
constants, due to division), and must have their plasticity set
to false.

These changes seem to correct a previous crash on default initial-
ization, likely due to a dependency on the Ne variable, which has
been removed.